# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fintech Core Engine API
  version: 0.1.3
  description: |
    Infraestructura financiera backend con ledger de doble entrada, wallets y APIs REST.

    Todos los endpoints (excepto `/health`) requieren JWT Bearer con scopes en el payload:
    ```json
    { "sub": "userId", "scopes": ["wallets:create", "transactions:write", ...] }
    ```
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: http://localhost:3000
    description: Health check (unversioned)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT firmado con `JWT_SECRET`. Payload debe incluir array `scopes`.
        Scopes disponibles: `tenants:admin`,
        `wallets:create`, `wallets:read`,
        `transactions:write`, `transactions:read`, `value:issue`, `value:redeem`,
        `compliance:read`, `compliance:write`,
        `integrations:read`, `integrations:write`.
  parameters:
    walletId:
      name: walletId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateWalletRequest:
      type: object
      required: [tenant_id, owner_id, currency]
      properties:
        tenant_id:
          type: string
          format: uuid
          example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        owner_id:
          type: string
          example: user-123
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: CLP
...
    WalletResponse:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        owner_id:
          type: string
        currency:
          type: string
          minLength: 3
          maxLength: 3
        status:
          type: string
          enum: [active, suspended, closed]
          example: active
    BalanceResponse:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        available:
          type: string
          description: Balance disponible derivado del ledger.
          example: "9500.00"
        ledger:
          type: string
          description: Balance contable del ledger (igual a available en esta versión).
          example: "9500.00"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: CLP
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Insufficient funds
        error:
          type: string
          example: Bad Request
paths:
...
  /wallets:
    post:
      summary: Crear wallet
      description: |
        Crea un wallet con su cuenta contable principal (liability) asociada en el ledger.
        Scope requerido: `wallets:create`
      tags: [Wallets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletRequest'
      responses:
        '201':
          description: Wallet creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Parámetros inválidos o faltantes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token JWT inválido o ausente
        '403':
          description: Scope `wallets:create` requerido
  /wallets/{walletId}:
    get:
      summary: Obtener wallet
      description: |
        Retorna el detalle de un wallet por su ID.
        Scope requerido: `wallets:read`
      tags: [Wallets]
      parameters:
        - $ref: '#/components/parameters/walletId'
      responses:
        '200':
          description: Detalle del wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '401':
          description: Token JWT inválido o ausente
        '403':
          description: Scope `wallets:read` requerido
        '404':
          description: Wallet no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wallets/{walletId}/balance:
    get:
      summary: Consultar balance de wallet
      description: |
        Balance derivado del ledger de doble entrada.
        El campo `available` refleja el saldo neto (créditos − débitos) de los asientos contables.
        Scope requerido: `wallets:read`
      tags: [Wallets]
      parameters:
        - $ref: '#/components/parameters/walletId'
      responses:
        '200':
          description: Balance del wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '401':
          description: Token JWT inválido o ausente
        '403':
          description: Scope `wallets:read` requerido
        '404':
          description: Wallet no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...